Hackathon_2/Todo app phase 5/backend/requirements.txt
# Core FastAPI dependencies
fastapi==0.109.0
uvicorn[standard]==0.27.0
python-multipart==0.0.18

# Database and ORM
sqlmodel==0.0.14
psycopg2-binary==2.9.9
asyncpg==0.29.0

# Data validation
pydantic>=2.8.0
pydantic-settings>=2.6.1

# Authentication and security
python-jose[cryptography]==3.4.0
passlib[bcrypt]==1.7.4
bcrypt==4.1.2
cryptography==42.0.0

# HTTP client
httpx>=0.27.0

# AI/LLM integration
openai>=1.40.0

# MCP (Model Context Protocol)
mcp>=1.0.0

# Environment variables
python-dotenv==1.0.0

# Background jobs and event streaming (Phase 6: Due Dates and Reminders)
aiokafka==0.10.0
apscheduler==3.10.4
